src/requirements.txt
alembic==0.8.7
blinker==1.4
boto==2.42.0
cffi==1.7.0
click==6.6
contextlib2==0.5.4
cryptography==1.4
decorator==4.0.10
elasticsearch==2.3.0
elasticsearch-dsl==2.0.0
enum34==1.1.6
Flask==1.0
Flask-Elasticsearch==0.2.5
Flask-Login==0.3.2
Flask-Mail==0.9.1
Flask-Migrate==2.0.0
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
Flask-WTF==0.12
frozendict==1.0
gcs-oauth2-boto-plugin==1.9
google-api-python-client==1.5.1
httplib2==0.9.2
idna==2.1
infinity==1.4
intervals==0.8.0
ipaddress==1.0.16
itsdangerous==0.24
Jinja2==2.11.3
Mako==1.0.4
MarkupSafe==0.23
oauth2client==1.5.2
Pillow==8.3.2
psycopg2==2.6.2
py==1.4.31
pyasn1==0.1.9
pyasn1-modules==0.0.8
pycparser==2.14
pyOpenSSL==17.5.0
pytest==3.0.3
pytest-flask==0.10.0
python-dateutil==2.5.3
python-editor==1.0.1
python-http-client==2.1.1
pytz==2016.6.1
raven==5.24.3
requests==2.11.1
retry-decorator==1.1.0
rsa==4.7
sendgrid==3.1.10
simplejson==3.8.2
six==1.10.0
smtpapi==0.3.1
SocksiPy-branch==1.1
SQLAlchemy==1.3.0
stripe==1.41.0
uritemplate==0.6
urllib3==1.26.5
validators==0.11.0
Werkzeug==0.11.9
WTForms==2.1
WTForms-Components==0.10.0
